move laminography methods to new module laminography.py
