+++ mbircone/__init__.py
@@ -4,4 +4,5 @@ from .cone3D import *
 from .mace import mace3D, mace4D
 from .phantom import *
 from .multinode import *
+from .laminography import *
 #__all__ = ['auto_sigma_x','auto_sigma_p','auto_sigma_y', 'calc_weights', 'compute_sino_params', 'compute_img_params', 'compute_img_size', 'project', 'recon', 'pad_roi2ror', 'extract_roi_from_ror']
